04/03

+++ src/main/java/com/daw/atm/models/Diposit.java
@@ -1,22 +1,22 @@
 package com.daw.atm.models;
 
 public class Diposit {
-    private double valor;
-    private double quantitat;
+    private int valor;
+    private int quantitat;
     
-    public double getValor() {
+    public int getValor() {
         return valor;
     }
     public void setValor(int valor) {
         this.valor = valor;
     }
-    public double getQuantitat() {
+    public int getQuantitat() {
         return quantitat;
     }
-    public void setQuantitat(double quantitat) {
+    public void setQuantitat(int quantitat) {
         this.quantitat = quantitat;
     }
-    public Diposit(double valor, double quantitat) {
+    public Diposit(int valor, int quantitat) {
         this.valor = valor;
         this.quantitat = quantitat;
     }
